refactor(projects): one component per file
Per Ayush's review on PR #1220 round 3 (cells.tsx:37 and cells.tsx:59):
each reusable component gets its own kebab-case file. Splits the former
cells.tsx (7 components aggregated) into seven per-component files:

- dot.tsx                  — risk indicator (cva variant)
- stages-icon.tsx          — phase donut (cva variant; @todo pointing
                             at frappe-ui-react#248 for the eventual
                             upstream Stages icon)
- date-cell.tsx            — Calendar icon + formatProjectDate
- project-name-cell.tsx    — risk dot + name + project-detail nav
- phase-cell.tsx           — stages icon + phase label
- budget-progress-cell.tsx — Progress with tier-based cva classes
- employee-cell.tsx        — Avatar + name + /desk/user/<email> link

cell.tsx now imports each cell component from its own file. cells.tsx
deleted. cva variant definitions stay in constants.ts as shared source
of truth for all three variant families.
